feat: migrate beneficiary module
