+++ features/beneficiary/presentation/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("invoicer.library")
+    id("invoicer.multiplatform.library")
     id("invoicer.compose")
     alias(libs.plugins.paparazzi)
 }
@@ -8,34 +8,45 @@ android {
     namespace = "io.github.alaksion.invoicer.features.beneficiary.presentation"
 }
 
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            // Compose
+            implementation(compose.ui)
+            implementation(compose.material3)
+            implementation(compose.components.resources)
+
+            // Koin
+            implementation(project.dependencies.platform(libs.koin.bom))
+            implementation(libs.koin.core)
+
+            // Voyager
+            implementation(libs.bundles.voyager)
+            implementation(libs.voyager.transitions)
+
+            // Kotlin
+            implementation(libs.immutable.collections)
+            implementation(libs.datetime)
+
+            // Foundation
+            implementation(projects.foundation.network)
+            implementation(projects.foundation.navigation)
+            implementation(projects.foundation.designSystem)
+            implementation(projects.foundation.ui)
+            implementation(projects.foundation.utils)
+
+            // Libs
+            implementation(projects.features.beneficiary.services)
+            implementation(projects.foundation.watchers)
+        }
+
+        androidMain.dependencies {
+            implementation(libs.androidx.activity.compose)
+        }
+    }
+}
+
 dependencies {
-    // Compose
-    implementation(libs.bundles.compose.ui)
-    implementation(libs.androidx.activity.compose)
-    debugImplementation(libs.bundles.compose.debug)
-
-    // Koin
-    implementation(platform(libs.koin.bom))
-    implementation(libs.koin.core)
-
-    // Voyager
-    implementation(libs.bundles.voyager)
-    implementation(libs.voyager.transitions)
-
-    // Kotlin
-    implementation(libs.immutable.collections)
-    implementation(libs.datetime)
-
-    // Foundation
-    implementation(projects.foundation.network)
-    implementation(projects.foundation.navigation)
-    implementation(projects.foundation.designSystem)
-    implementation(projects.foundation.ui)
-    implementation(projects.foundation.utils)
-
-    // Feature
-    implementation(projects.features.beneficiary.services)
-    implementation(projects.foundation.watchers)
 
     // Test
     testImplementation(kotlin("test"))
