+++ features/beneficiary/presentation/src/commonMain/kotlin/io/github/alaksion/invoicer/features/beneficiary/presentation/navigation/BeneficiaryScreenModule.kt
