+++ features/beneficiary/presentation/src/commonMain/kotlin/io/github/alaksion/invoicer/features/beneficiary/presentation/screen/create/steps/BeneficiaryBankInfoStep.kt
@@ -1,0 +1,163 @@
+package io.github.alaksion.invoicer.features.beneficiary.presentation.screen.create.steps
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.FocusRequester.Companion.FocusRequesterFactory.component1
+import androidx.compose.ui.focus.FocusRequester.Companion.FocusRequesterFactory.component2
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.input.ImeAction
+import cafe.adriel.voyager.core.screen.Screen
+import cafe.adriel.voyager.koin.koinNavigatorScreenModel
+import cafe.adriel.voyager.navigator.LocalNavigator
+import cafe.adriel.voyager.navigator.currentOrThrow
+import invoicer.features.beneficiary.presentation.generated.resources.Res
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_address_label
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_address_placeholder
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_info_subtitle
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_info_title
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_name_label
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_bank_name_placeholder
+import invoicer.features.beneficiary.presentation.generated.resources.create_beneficiary_continue_cta
+import io.github.alaksion.invoicer.features.beneficiary.presentation.screen.create.CreateBeneficiaryScreenModel
+import io.github.alaksion.invoicer.foundation.designSystem.components.InputField
+import io.github.alaksion.invoicer.foundation.designSystem.components.ScreenTitle
+import io.github.alaksion.invoicer.foundation.designSystem.components.buttons.BackButton
+import io.github.alaksion.invoicer.foundation.designSystem.components.buttons.PrimaryButton
+import io.github.alaksion.invoicer.foundation.designSystem.components.spacer.SpacerSize
+import io.github.alaksion.invoicer.foundation.designSystem.components.spacer.VerticalSpacer
+import io.github.alaksion.invoicer.foundation.designSystem.tokens.Spacing
+import org.jetbrains.compose.resources.stringResource
+
+internal class BeneficiaryBankInfoStep : Screen {
+    @Composable
+    override fun Content() {
+        val navigator = LocalNavigator.currentOrThrow
+        val screenModel = navigator.koinNavigatorScreenModel<CreateBeneficiaryScreenModel>()
+        val state by screenModel.state.collectAsState()
+
+        StateContent(
+            address = state.bankAddress,
+            bankName = state.bankName,
+            onAddressChange = screenModel::updateBankAddress,
+            onBankNameChange = screenModel::updateBankName,
+            onBack = { navigator.pop() },
+            onContinue = { navigator.push(BeneficiaryConfirmationStep()) },
+            buttonEnabled = state.bankInfoIsValid
+        )
+    }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Composable
+    fun StateContent(
+        address: String,
+        bankName: String,
+        buttonEnabled: Boolean,
+        onAddressChange: (String) -> Unit,
+        onBankNameChange: (String) -> Unit,
+        onBack: () -> Unit,
+        onContinue: () -> Unit,
+    ) {
+        val (nameFocus, addressFocus) = FocusRequester.createRefs()
+        val keyboard = LocalSoftwareKeyboardController.current
+
+        Scaffold(
+            modifier = Modifier.imePadding(),
+            topBar = {
+                TopAppBar(
+                    title = {},
+                    navigationIcon = {
+                        BackButton(onBackClick = onBack)
+                    }
+                )
+            },
+            bottomBar = {
+                PrimaryButton(
+                    label = stringResource(Res.string.create_beneficiary_continue_cta),
+                    onClick = {
+                        keyboard?.hide()
+                        onContinue()
+                    },
+                    isEnabled = buttonEnabled,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(Spacing.medium),
+                )
+            }
+        ) { scaffoldPadding ->
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(Spacing.medium)
+                    .padding(scaffoldPadding)
+            ) {
+                ScreenTitle(
+                    title = stringResource(Res.string.create_beneficiary_bank_info_title),
+                    subTitle = stringResource(Res.string.create_beneficiary_bank_info_subtitle)
+                )
+                VerticalSpacer(SpacerSize.XLarge3)
+                InputField(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .focusRequester(nameFocus),
+                    value = bankName,
+                    onValueChange = onBankNameChange,
+                    keyboardOptions = KeyboardOptions(
+                        imeAction = ImeAction.Next
+                    ),
+                    keyboardActions = KeyboardActions(
+                        onNext = { addressFocus.requestFocus() }
+                    ),
+                    label = {
+                        Text(
+                            text = stringResource(Res.string.create_beneficiary_bank_name_label)
+                        )
+                    },
+                    placeholder = {
+                        Text(
+                            text = stringResource(Res.string.create_beneficiary_bank_name_placeholder)
+                        )
+                    }
+                )
+                VerticalSpacer(SpacerSize.Medium)
+                InputField(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .focusRequester(addressFocus),
+                    value = address,
+                    onValueChange = onAddressChange,
+                    keyboardOptions = KeyboardOptions(
+                        imeAction = ImeAction.Done
+                    ),
+                    keyboardActions = KeyboardActions(
+                        onDone = { keyboard?.hide() }
+                    ),
+                    label = {
+                        Text(
+                            text = stringResource(Res.string.create_beneficiary_bank_address_label)
+                        )
+                    },
+                    placeholder = {
+                        Text(
+                            text = stringResource(Res.string.create_beneficiary_bank_address_placeholder)
+                        )
+                    }
+                )
+            }
+        }
+    }
+}
