--- features/beneficiary/presentation/src/main/kotlin/io/github/alaksion/invoicer/features/beneficiary/presentation/screen/create/CreateBeneficiaryFlow.kt
@@ -1,15 +1,0 @@
-package io.github.alaksion.invoicer.features.beneficiary.presentation.screen.create
-
-import androidx.compose.runtime.Composable
-import cafe.adriel.voyager.core.screen.Screen
-import cafe.adriel.voyager.navigator.Navigator
-import cafe.adriel.voyager.transitions.SlideTransition
-import io.github.alaksion.invoicer.features.beneficiary.presentation.screen.create.steps.BeneficiaryNameStep
-
-internal class CreateBeneficiaryFlow : Screen {
-
-    @Composable
-    override fun Content() {
-        Navigator(BeneficiaryNameStep()) { SlideTransition(it) }
-    }
-}
