--- features/beneficiary/presentation/src/main/kotlin/io/github/alaksion/invoicer/features/beneficiary/presentation/screen/list/components/BeneficiaryCard.kt
@@ -1,62 +1,0 @@
-package io.github.alaksion.invoicer.features.beneficiary.presentation.screen.list.components
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Business
-import androidx.compose.material.icons.outlined.ChevronRight
-import androidx.compose.material3.Card
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import io.github.alaksion.invoicer.features.beneficiary.presentation.R
-import io.github.alaksion.invoicer.features.beneficiary.services.domain.model.BeneficiaryModel
-import io.github.alaksion.invoicer.foundation.designSystem.tokens.Spacing
-
-@Composable
-internal fun BeneficiaryCard(
-    data: BeneficiaryModel,
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = modifier.clickable(onClick = onClick)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(Spacing.small),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(Spacing.medium)
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Business,
-                contentDescription = null
-            )
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = data.name,
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.SemiBold
-                )
-                Text(
-                    text = stringResource(R.string.beneficiary_list_item_swift, data.swift),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-            Icon(
-                imageVector = Icons.Outlined.ChevronRight,
-                contentDescription = null
-            )
-        }
-    }
-}
