--- foundation/exception/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    id("invoicer.library")
-}
-
-android {
-    namespace = "foundation.exception"
-}
